requirements.txt
# Core Django
Django==5.1.15
gunicorn==22.0.0

# Database
mysqlclient==2.2.4

# Image Processing (für PWA Icons)
Pillow==12.1.0

# Environment Variables
python-dotenv==1.0.1

# PDF Export (Hybrid: WeasyPrint für Linux, xhtml2pdf als Fallback)
WeasyPrint==68.0
xhtml2pdf==0.2.17
matplotlib==3.10.8
cairosvg==2.7.1  # SVG zu PNG Rendering für Body-Map
qrcode[pil]==8.0  # QR-Code Generation für Plan-PDF Export

# AI Coach - LLM Integration
ollama==0.4.3
openai==1.58.1
keyring==25.5.0

# Optional: Redis für Caching & Sessions
redis==5.0.1
django-redis==5.4.0

# Web Push Notifications
pywebpush==1.14.1

# ML Coach - Machine Learning (CPU-only, keine GPU)
scikit-learn==1.6.1
joblib==1.5.0

# Testing & Quality Assurance
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0

# Monitoring & Error Tracking
sentry-sdk==2.22.0
django-axes==7.1.0  # Brute-force protection

# Code Quality & Formatting
black==25.1.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
django-stubs==5.1.2
pre-commit==4.0.1
